Extract cmd filter impl to own file + move test
